requirements.txt
# ============================================================
# Face Recognition API - Requirements
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# ----------------------------------------------------------
# Deep Learning Framework
# ----------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0

# ----------------------------------------------------------
# Model & Face Detection
# ----------------------------------------------------------
transformers>=4.30.0
facenet-pytorch>=2.5.0

# ----------------------------------------------------------
# Web Framework
# ----------------------------------------------------------
flask>=2.3.0
flask-cors>=4.0.0

# ----------------------------------------------------------
# Image Processing
# ----------------------------------------------------------
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ----------------------------------------------------------
# Utilities
# ----------------------------------------------------------
psutil>=5.9.0

# ----------------------------------------------------------
# Training Only (Optional)
# ----------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
roboflow>=1.1.0
